requirements.txt
flask
requests
gunicorn
google-api-python-client
google-auth
google-auth-httplib2
google-auth-oauthlib
supabase
